+++ .gitignore
@@ -1,2 +1,16 @@
-bin
-dist
+*.class
+*.DS_store
+.classpath
+.project
+.metadata
+.settings
+.settings/*
+.idea
+/*/target/*
+/target/
+/target
+*.log
+*.log.gz
+.idea/*
+*.iml
+*.patch
