--- build.xml
@@ -1,1 +1,0 @@
-build_with_tests.xml
