+++ src/main/java/com/alchemyapi/api/AlchemyAPI_ImageParams.java
