+++ src/test/java/com/alchemyapi/CategoryTest.java
@@ -1,39 +1,51 @@
-package com.alchemyapi.test;
+package com.alchemyapi;
 
 import com.alchemyapi.api.AlchemyAPI;
-
-import org.xml.sax.SAXException;
+import com.alchemyapi.api.AlchemyAPI_CategoryParams;
+import com.alchemyapi.api.AlchemyAPI_Params;
 import org.w3c.dom.Document;
-import java.io.*;
+import org.xml.sax.SAXException;
+
 import javax.xml.parsers.ParserConfigurationException;
-import javax.xml.xpath.XPathExpressionException;
 import javax.xml.transform.Transformer;
 import javax.xml.transform.TransformerException;
 import javax.xml.transform.TransformerFactory;
 import javax.xml.transform.dom.DOMSource;
 import javax.xml.transform.stream.StreamResult;
+import javax.xml.xpath.XPathExpressionException;
+import java.io.BufferedReader;
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.FileReader;
+import java.io.IOException;
+import java.io.StringWriter;
 
-class KeywordTest {
-    public static void main(String[] args) throws IOException, SAXException,
-            ParserConfigurationException, XPathExpressionException {
+class CategoryTest {
+    public static void main(String[] args)
+        throws IOException, SAXException,
+               ParserConfigurationException, XPathExpressionException
+    {
         // Create an AlchemyAPI object.
         AlchemyAPI alchemyObj = AlchemyAPI.GetInstanceFromFile("api_key.txt");
 
-        // Extract topic keywords for a web URL.
-        Document doc = alchemyObj.URLGetRankedKeywords("http://www.techcrunch.com/");
+        // Categorize a web URL by topic.
+        Document doc = alchemyObj.URLGetCategory("http://www.techcrunch.com/");
         System.out.println(getStringFromDocument(doc));
 
-        // Extract topic keywords for a text string.
-        doc = alchemyObj.TextGetRankedKeywords(
-            "Hello there, my name is Bob Jones.  I live in the United States of America.  " +
-            "Where do you live, Fred?");
+        // Categorize some text.
+        doc = alchemyObj.TextGetCategory("Latest on the War in Iraq.");
         System.out.println(getStringFromDocument(doc));
 
         // Load a HTML document to analyze.
         String htmlDoc = getFileContents("data/example.html");
 
-        // Extract topic keywords for a HTML document.
-        doc = alchemyObj.HTMLGetRankedKeywords(htmlDoc, "http://www.test.com/");
+        // Categorize a HTML document by topic.
+        doc = alchemyObj.HTMLGetCategory(htmlDoc, "http://www.test.com/");
+        System.out.println(getStringFromDocument(doc));
+        
+        AlchemyAPI_CategoryParams categoryParams = new AlchemyAPI_CategoryParams();
+        categoryParams.setOutputMode(AlchemyAPI_Params.OUTPUT_RDF);
+        doc = alchemyObj.HTMLGetCategory(htmlDoc, "http://www.test.com/", categoryParams);
         System.out.println(getStringFromDocument(doc));
     }
