+++ src/test/java/com/alchemyapi/ConceptTest.java
@@ -1,4 +1,4 @@
-package com.alchemyapi.test;
+package com.alchemyapi;
 
 import com.alchemyapi.api.AlchemyAPI;
 
@@ -13,23 +13,26 @@ import javax.xml.transform.TransformerFactory;
 import javax.xml.transform.dom.DOMSource;
 import javax.xml.transform.stream.StreamResult;
 
-class FeedLinksTest {
-    public static void main(String[] args)
-        throws IOException, SAXException,
-               ParserConfigurationException, XPathExpressionException
-    {
+class ConceptTest {
+    public static void main(String[] args) throws IOException, SAXException,
+            ParserConfigurationException, XPathExpressionException {
         // Create an AlchemyAPI object.
         AlchemyAPI alchemyObj = AlchemyAPI.GetInstanceFromFile("api_key.txt");
 
-        // Extract RSS / ATOM feed links from a web URL.
-        Document doc = alchemyObj.URLGetFeedLinks("http://www.techcrunch.com/");
+        // Extract concept tags for a web URL.
+        Document doc = alchemyObj.URLGetRankedConcepts("http://www.techcrunch.com/");
+        System.out.println(getStringFromDocument(doc));
+
+        // Extract concept tags for a text string.
+        doc = alchemyObj.TextGetRankedConcepts(
+            "This thing has a steering wheel, tires, and an engine.  Do you know what it is?");
         System.out.println(getStringFromDocument(doc));
 
         // Load a HTML document to analyze.
         String htmlDoc = getFileContents("data/example.html");
 
-        // Extract RSS / ATOM feed links from a HTML document.
-        doc = alchemyObj.HTMLGetFeedLinks(htmlDoc, "http://www.test.com/");
+        // Extract concept tags for a HTML document.
+        doc = alchemyObj.HTMLGetRankedConcepts(htmlDoc, "http://www.test.com/");
         System.out.println(getStringFromDocument(doc));
     }
