+++ src/test/java/com/alchemyapi/MicroformatsTest.java
@@ -1,19 +1,24 @@
-package com.alchemyapi.test;
+package com.alchemyapi;
 
-import com.alchemyapi.api.*;
-
-import org.xml.sax.SAXException;
+import com.alchemyapi.api.AlchemyAPI;
 import org.w3c.dom.Document;
-import java.io.*;
+import org.xml.sax.SAXException;
+
 import javax.xml.parsers.ParserConfigurationException;
-import javax.xml.xpath.XPathExpressionException;
 import javax.xml.transform.Transformer;
 import javax.xml.transform.TransformerException;
 import javax.xml.transform.TransformerFactory;
 import javax.xml.transform.dom.DOMSource;
 import javax.xml.transform.stream.StreamResult;
+import javax.xml.xpath.XPathExpressionException;
+import java.io.BufferedReader;
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.FileReader;
+import java.io.IOException;
+import java.io.StringWriter;
 
-class TaxonomyTest {
+public class MicroformatsTest {
     public static void main(String[] args)
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
@@ -21,21 +26,15 @@ class TaxonomyTest {
         // Create an AlchemyAPI object.
         AlchemyAPI alchemyObj = AlchemyAPI.GetInstanceFromFile("api_key.txt");
 
-        // Extract a ranked list of relations for a web URL.
-        Document doc = alchemyObj.URLGetTaxonomy("http://www.techcrunch.com/");
-        System.out.println(getStringFromDocument(doc));
-
-        // Extract a ranked taxonomy from a text string.
-        doc = alchemyObj.TextGetTaxonomy(
-            "Hello there, my name is Bob Jones.  I live in the United States of America.  " +
-            "Where do you live, Fred?");
+        // Extract microformats data from a web URL.
+        Document doc = alchemyObj.URLGetMicroformats("http://microformats.org/wiki/hcard");
         System.out.println(getStringFromDocument(doc));
 
         // Load a HTML document to analyze.
-        String htmlDoc = getFileContents("data/example.html");
+        String htmlDoc = getFileContents("data/microformats.html");
 
-        // Extract a ranked taxonomy from a HTML document.
-        doc = alchemyObj.HTMLGetTaxonomy(htmlDoc, "http://www.test.com/");
+        // Extract microformats data from a HTML document.
+        doc = alchemyObj.HTMLGetMicroformats(htmlDoc, "http://www.test.com/");
         System.out.println(getStringFromDocument(doc));
     }
