+++ src/test/java/com/alchemyapi/ParameterTest.java
@@ -1,4 +1,4 @@
-package com.alchemyapi.test;
+package com.alchemyapi;
 
 import com.alchemyapi.api.*;
