+++ src/test/java/com/alchemyapi/SentimentTest.java
@@ -1,4 +1,4 @@
-package com.alchemyapi.test;
+package com.alchemyapi;
 
 import com.alchemyapi.api.AlchemyAPI;
 import com.alchemyapi.api.*;
