+++ src/test/java/com/alchemyapi/TaxonomyTest.java
@@ -1,4 +1,4 @@
-package com.alchemyapi.test;
+package com.alchemyapi;
 
 import com.alchemyapi.api.*;
 
@@ -13,7 +13,7 @@ import javax.xml.transform.TransformerFactory;
 import javax.xml.transform.dom.DOMSource;
 import javax.xml.transform.stream.StreamResult;
 
-class CombinedTest {
+class TaxonomyTest {
     public static void main(String[] args)
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
@@ -21,22 +21,21 @@ class CombinedTest {
         // Create an AlchemyAPI object.
         AlchemyAPI alchemyObj = AlchemyAPI.GetInstanceFromFile("api_key.txt");
 
-        // Extract combined data for a web URL.
-        Document doc = alchemyObj.URLGetCombined("http://www.techcrunch.com/");
+        // Extract a ranked list of relations for a web URL.
+        Document doc = alchemyObj.URLGetTaxonomy("http://www.techcrunch.com/");
         System.out.println(getStringFromDocument(doc));
 
-        // Extract combined data from a text string.
-        doc = alchemyObj.TextGetCombined(
+        // Extract a ranked taxonomy from a text string.
+        doc = alchemyObj.TextGetTaxonomy(
             "Hello there, my name is Bob Jones.  I live in the United States of America.  " +
             "Where do you live, Fred?");
         System.out.println(getStringFromDocument(doc));
 
-	// Only extract entities & keywords
-	AlchemyAPI_CombinedParams combinedParams = new AlchemyAPI_CombinedParams();
-	combinedParams.setSentiment(true);
-	combinedParams.setExtract("entity");
-	combinedParams.setExtract("keyword");
-	doc = alchemyObj.TextGetCombined("Madonna enjoys tasty Pepsi.  I love her style.", combinedParams);
+        // Load a HTML document to analyze.
+        String htmlDoc = getFileContents("data/example.html");
+
+        // Extract a ranked taxonomy from a HTML document.
+        doc = alchemyObj.HTMLGetTaxonomy(htmlDoc, "http://www.test.com/");
         System.out.println(getStringFromDocument(doc));
     }
