+++ src/test/resources/api_key.txt
@@ -1,0 +1,1 @@
+<INSERT_API_KEY>
