+++ src/test/resources/data/cat.jpg
